criação de testes unitarios em alguns dos serviços

+++ src/main/java/com/graphql/project/dtos/CreateCustomer.java
@@ -10,5 +10,5 @@ public record CreateCustomer(Integer salesRepEmployeeNum,
                              String address2, String city,
                              String state, int postalCode,
                              String country,
-                             BigDecimal creditLimit) {
+                             int creditLimit) {
 }
